requirements.txt
notion-client
python-dotenv
pyyaml
beautifulsoup4
requests
